add a sqlite interface to use in slips main db and the trust db
